Add necessary requirements for moving clear database into settings
